Done manager and canvas, now need make tool logic

+++ include/Vec2.hpp
@@ -1,27 +1,28 @@
 #pragma once
 
+#include <cstdint>
 #include <cstdio>
 #include <iostream>
 
 class Vec2Kernel
 {
   private:
-    size_t x_, y_;
+    int_fast32_t x_, y_;
 
   public:
-    Vec2Kernel(size_t x = 0, size_t y = 0) : x_(x), y_(y) {}
+    Vec2Kernel(int_fast32_t x = 0, int_fast32_t y = 0) : x_(x), y_(y) {}
 
-    size_t getX() const { return x_; }
-    size_t getY() const { return y_; }
+    int_fast32_t getX() const { return x_; }
+    int_fast32_t getY() const { return y_; }
 
-    size_t setX(size_t x);
-    size_t setY(size_t y);
+    int_fast32_t setX(int_fast32_t x);
+    int_fast32_t setY(int_fast32_t y);
 
     const Vec2Kernel &operator=(const Vec2Kernel &ker);
 
     const Vec2Kernel &operator+=(const Vec2Kernel &arg);
     const Vec2Kernel &operator-=(const Vec2Kernel &arg);
-    const Vec2Kernel &operator*=(size_t coef);
+    const Vec2Kernel &operator*=(int_fast32_t coef);
 };
 
 class Vec2
@@ -30,16 +31,16 @@ class Vec2
     Vec2Kernel Kernel;
 
   public:
-    Vec2(size_t x = 0, size_t y = 0) : Kernel(x, y) {}
+    Vec2(int_fast32_t x = 0, int_fast32_t y = 0) : Kernel(x, y) {}
 
     Vec2(const Vec2 &&vec) { Kernel = vec.Kernel; }
     Vec2(const Vec2 &vec) { Kernel = vec.Kernel; }
 
-    size_t getX() const { return Kernel.getX(); }
-    size_t getY() const { return Kernel.getY(); }
+    int_fast32_t getX() const { return Kernel.getX(); }
+    int_fast32_t getY() const { return Kernel.getY(); }
 
-    size_t setX(size_t x) { return Kernel.setX(x); }
-    size_t setY(size_t y) { return Kernel.setY(y); }
+    int_fast32_t setX(int_fast32_t x) { return Kernel.setX(x); }
+    int_fast32_t setY(int_fast32_t y) { return Kernel.setY(y); }
 
     friend Vec2          operator*(double coef, const Vec2 &vec);
     friend std::ostream &operator<<(std::ostream &os, Vec2 &vec);
@@ -49,8 +50,8 @@ class Vec2
     Vec2        operator+(const Vec2 &vec) const;
     Vec2        operator-(const Vec2 &vec) const;
     Vec2        operator%(const Vec2 &vec) const;
-    Vec2        operator*(size_t coef) const;
-    Vec2        operator/(size_t coef) const { return this->operator*(1 / coef); }
+    Vec2        operator*(int_fast32_t coef) const;
+    Vec2        operator/(int_fast32_t coef) const { return this->operator*(1 / coef); }
     Vec2        operator-() const { return (-1) * (*this); }
 
     bool operator==(const Vec2 &vec) const;
@@ -60,7 +61,7 @@ class Vec2
 
     void operator+=(const Vec2 &arg) { Kernel += arg.Kernel; }
     void operator-=(const Vec2 &arg) { Kernel -= arg.Kernel; }
-    void operator*=(size_t coef) { Kernel *= coef; }
+    void operator*=(int_fast32_t coef) { Kernel *= coef; }
 };
 
 inline const Vec2Kernel &Vec2Kernel::operator=(const Vec2Kernel &ker)
@@ -87,7 +88,7 @@ inline const Vec2Kernel &Vec2Kernel::operator-=(const Vec2Kernel &ker)
     return *this;
 }
 
-inline const Vec2Kernel &Vec2Kernel::operator*=(size_t coef)
+inline const Vec2Kernel &Vec2Kernel::operator*=(int_fast32_t coef)
 {
     x_ *= coef;
     y_ *= coef;
@@ -107,7 +108,7 @@ inline const Vec2 &Vec2::operator=(const Vec2 &vec)
     return *this;
 }
 
-inline Vec2 Vec2::operator*(size_t coef) const
+inline Vec2 Vec2::operator*(int_fast32_t coef) const
 {
     return Vec2(Kernel.getX() * coef, Kernel.getY() * coef);
 }
